Move all non-test devDependencies to dependencies
I made a mistake.
I had all these dependencies in dev
because that worked when all the code got bundled up via Webpack.
But now that this only Babelifies the consumer
has to somehow get the dependencies that this code refers to.

The old setup worked locally because installing a package locally
exposes the consumer to that library's entire node_modules,
but once it gets installed via NPM or GitHub
only actual dependencies are installed. Whoops.